+++ package.json
@@ -10,6 +10,7 @@
     "deploy": "npx projen deploy",
     "destroy": "npx projen destroy",
     "diff": "npx projen diff",
+    "eject": "npx projen eject",
     "eslint": "npx projen eslint",
     "package": "npx projen package",
     "post-compile": "npx projen post-compile",
@@ -34,7 +35,7 @@
     "@types/node": "^12",
     "@typescript-eslint/eslint-plugin": "^5",
     "@typescript-eslint/parser": "^5",
-    "aws-cdk": "^2.8.0",
+    "aws-cdk": "^2.15.0",
     "esbuild": "^0.14.13",
     "eslint": "^8",
     "eslint-import-resolver-node": "^0.3.6",
@@ -44,13 +45,13 @@
     "jest-junit": "^13",
     "json-schema": "^0.4.0",
     "npm-check-updates": "^12",
-    "projen": "^0.51.14",
+    "projen": "^0.52.50",
     "ts-jest": "^27.1.3",
     "ts-node": "^9",
     "typescript": "^4.5.5"
   },
   "dependencies": {
-    "aws-cdk-lib": "^2.8.0",
+    "aws-cdk-lib": "^2.15.0",
     "constructs": "^10.0.5"
   },
   "keywords": [
