+++ src/main.ts
@@ -1,19 +1,5 @@
-import { App, Stack, StackProps } from 'aws-cdk-lib';
-import * as lambda from 'aws-cdk-lib/aws-lambda';
-import { Construct } from 'constructs';
-
-export class LambdaStack extends Stack {
-  constructor(scope: Construct, id: string, props: StackProps = {}) {
-    super(scope, id, props);
-
-    new lambda.Function(this, 'ExampleFunction', {
-      functionName: 'example-lambda',
-      code: lambda.Code.fromAsset('lambda'),
-      handler: 'hello.handler',
-      runtime: lambda.Runtime.NODEJS_14_X,
-    });
-  }
-}
+import { App } from 'aws-cdk-lib';
+import { CdkPipelineStack } from './cdk-pipeline-stack';
 
 // for development, use account/region from cdk cli
 const devEnv = {
@@ -23,7 +9,6 @@ const devEnv = {
 
 const app = new App();
 
-new LambdaStack(app, 'lambda-stack-dev', { env: devEnv });
-// new LambdaStack(app, 'lambda-stack-prod', { env: prodEnv });
+new CdkPipelineStack(app, 'CdkPipelineStack', { env: devEnv });
 
 app.synth();
